orogen: define associate_ports_to_transform
Unlike this method defines the port-to-transform association,
without assuming anything about the nature of the data the
port transports. This is unlike #transform_input and #transform_output.
Even more so as the former assumes that the data fed into the port
will be passed to the underlying transformer implementation.

The association is declared automatically by transform_input/
transform_output

each_transform_port now enumerates the ports declared by the new
method. This is consistent with the use of each_transform_port in
the rest of the code. Moreover, #find_transform_of_port also
resolves ports declared by the new method for the same reasons.